+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 -   Fixed an issue where CasualOS didn't support return type annotations on arrow functions.
 -   Fixed an issue where CasualOS didn't support generic type parameters on arrow functions.
+-   Fixed an issue where CasualOS didn't support "satisfies" TypeScript expressions.
 
 ## V4.2.1