+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 -   Fixed an issue where CasualOS didn't support generic type parameters on arrow functions.
 -   Fixed an issue where CasualOS didn't support "satisfies" TypeScript expressions.
 -   Fixed an issue where CasualOS didn't support some parenthesized expressions.
+-   Fixed an issue where the `value` attribute wouldn't work for `<select>` elements in custom apps.
 
 ## V4.2.1